Adding ability to remove the source file
When similar plugins convert the data, they remove the source.  For
instance, the markdown plugin will convert `*.md` into `*.html` and also
remove `*.md` so it never appears in the build directory.

This patch preserves the original behavior and adds an option that
removes the source `*.less` files if desired.

What was needed before this change:

    .use(less())
    .use(moveRemove({
        remove: [ ".*\\.less$" ]
    }))

With this commit added, you only need to do this:

    .use(less({
        removeSource: true
    }))